Refactor

+++ src/Sieve.cpp
@@ -442,7 +442,7 @@ void Sieve::cross_off_count(uint64_t prime, uint64_t i)
   count_ = 0;
   Wheel& wheel = wheel_[i];
   ASSERT(wheel.index <= 63);
-  uint32_t r = wheel.index >> 3;
+  uint64_t r = wheel.index >> 3;
   prime /= 30;
 
   const Array<uint8_t, 8> bitmask = {
@@ -485,7 +485,7 @@ void Sieve::cross_off_count(uint64_t prime, uint64_t i)
   }
 
   uint64_t m = wheel.multiple;
-  uint32_t s = wheel.index & 7;
+  uint64_t s = wheel.index & 7;
   uint64_t sieve_size = sieve_.size();
   uint8_t* sieve = &sieve_[0];
   uint64_t total_count = total_count_;
